justgram node add mvc pattern(createUser function)
