+++ models/createUserDao.js
@@ -1,0 +1,28 @@
+const { DataSource } = require("typeorm");
+
+const myDataSource = new DataSource({
+  type: process.env.TYPEORM_CONNECTION,
+  host: process.env.TYPEORM_HOST,
+  port: process.env.TYPEORM_PORT,
+  username: process.env.TYPEORM_USERNAME,
+  password: process.env.TYPEORM_PASSWORD,
+  database: process.env.TYPEORM_DATABASE,
+});
+
+myDataSource.initialize().then(() => {
+  console.log("Data Source has been initialized!");
+});
+
+const createUser = async (email, nickname, hashedPw, profile_image) => {
+  console.log("i am DAO 1");
+  await myDataSource.query(
+    `INSERT INTO users ( email, nickname, password, profile_image)
+    VALUES (
+      "${email}", "${nickname}", "${hashedPw}", "${profile_image}"
+      )`
+  );
+};
+
+module.exports = {
+  createUser,
+};
